Rename edit form component files
