+++ src/components/resource/resource-edit.js
